pyproject.toml
[tool.poetry]
name = "code-test-generator"
version = "0.1.0"
description = "Simple CLI app to generate code and tests"
authors = ["Chetan Choudhary"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.3"
langchain = "^0.2.7"
openai = "^1.35.13"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
